Fix importing package into cmake, rearrange dataset loader helpers
