+++ .github/workflows/ci.yml
@@ -22,6 +22,8 @@ jobs:
       name: Checkout
     - uses: jidicula/clang-format-action@v4.14.0
       name: Clang-Format
+      with:
+        clang-format-version: '10
 
   # run tests
   test:
