+++ python/evalio/datasets/enwide.py
@@ -3,25 +3,23 @@ import urllib.request
 from pathlib import Path
 from enum import auto
 
-from evalio.datasets.iterators import (
+from evalio.datasets.loaders import (
     LidarDensity,
     LidarFormatParams,
     LidarMajor,
     LidarPointStamp,
     LidarStamp,
     RosbagIter,
+    load_pose_csv,
 )
-from evalio.types import Trajectory
+from evalio.types import Trajectory, SE3, SO3
 import numpy as np
 from tqdm import tqdm
 
 from .base import (
-    SE3,
-    SO3,
     Dataset,
     ImuParams,
     LidarParams,
-    load_pose_csv,
     DatasetIterator,
 )
