+++ python/evalio/datasets/hilti_2022.py
@@ -3,25 +3,24 @@ import urllib
 import urllib.request
 from pathlib import Path
 
-from evalio.datasets.iterators import (
+from evalio.datasets.loaders import (
     LidarDensity,
     LidarFormatParams,
     LidarMajor,
     LidarPointStamp,
     LidarStamp,
     RosbagIter,
+    load_pose_csv,
 )
-from evalio.types import Trajectory
+from evalio.types import Trajectory, SO3
 import numpy as np
 from tqdm import tqdm
 
 from .base import (
     SE3,
-    SO3,
     Dataset,
     ImuParams,
     LidarParams,
-    load_pose_csv,
     DatasetIterator,
 )
