+++ python/evalio/types.py
@@ -1,5 +1,13 @@
-from ._cpp.types import *  # type: ignore  # noqa: F403
-from ._cpp.types import SE3, Stamp  # type: ignore
+from ._cpp.types import (  # type: ignore
+    SE3,
+    SO3,
+    Stamp,
+    Point,
+    LidarMeasurement,
+    LidarParams,
+    ImuParams,
+    ImuMeasurement,
+)
 from dataclasses import dataclass
 
 
@@ -21,3 +29,16 @@ class Trajectory:
     def transform_in_place(self, T: SE3):
         for i in range(len(self.poses)):
             self.poses[i] = self.poses[i] * T
+
+
+__all__ = [
+    "ImuMeasurement",
+    "ImuParams",
+    "LidarMeasurement",
+    "LidarParams",
+    "Point",
+    "SO3",
+    "SE3",
+    "Stamp",
+    "Trajectory",
+]
